// model/deezerReturnTypes/deezerArtist.ts
import { DeezerEntityType } from "./deezerEntityType.js";

export interface DeezerArtist {
  id: number;
  name: string;
  link: string;
  share: string;
  picture: string;
  picture_small: string;
  picture_medium: string;
  picture_big: string;
  picture_xl: string;
  nb_album: number;
  nb_fan: number;
  radio: boolean;
  tracklist: string;
  type: DeezerEntityType;
}